pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gateway"
version = "0.0.0"
description = "any-llm gateway service"
requires-python = ">=3.13"
dependencies = [
    "any-llm-sdk[all]",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "click>=8.1.0",
    "fastapi>=0.115.0",
    "prometheus-client>=0.20.0",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "mypy",
    "pytest>=8,<10",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.2.1,<8",
    "pytest-rerunfailures>=16.0",
    "pytest-timeout",
    "pytest-xdist>=3.6.1",
    "ruff>=0.15.9",
    "testcontainers[postgres]>=4.0.0",
]

[project.scripts]
gateway = "gateway.cli:main"
any-llm-gateway = "gateway.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 120
src = ["src", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"scripts/generate_openapi.py" = ["E402"]
"tests/integration/conftest.py" = ["E402"]

[tool.mypy]
python_version = "3.13"
strict = true
files = ["src", "tests", "scripts"]

[[tool.mypy.overrides]]
module = ["testcontainers.*"]
ignore_missing_imports = true
